<1.b> feat: use moment to print the date and time
In this commit we demonstrate how a pure ESM package from npm can be
consumed as a browser module.

Note that the majority of npm packages are not ESM ready, but there's an
ongoing effort to make that happen:

https://medium.com/@joeldenning/an-esm-bundle-for-any-npm-package-5f850db0e04d

Also, the more popular projects (like Angular, React, ...) are doing it
properly, without any need for third-party adaptations.

Note also that we need to refer to `./node_modules/@esm-bundle/moment` to
consume the npm package. This will be changed in the next commit to
'moment' leveraging import maps support in the portal, to make the mapping
between the two formats automagic.