<2> refactor: consume @esm-bundle/moment using an import map
For now, we will bundle a hand made importmap.json file inside the JAR so
that the server can process it to make the global import maps.

In the future, we may compute this file with the build script, by visiting
the JS modules starting from the entry points of the application (for now
just the implicit index.js file) so that everything happens automatically.

However, since the import maps technique alone will only work if all used
npm packages are ESM, it's not clear how that automagic will be
implemented in the end.