<3> feat: consume sbd package using a cjs: import map
In this commit, we move the importmap.json file to the root of the project.
The build will copy it to the proper place inside the JAR file
(META-INF/resources for now, though we may change it in the future).

Additionally, we add an import pointing to an URI with scheme cjs, like:

            "sbd": "cjs:./node_modules/sbd/lib/tokenizer.js"

This will be bundled using webpack and re-exported using ESM syntax so
that it can be loaded from the pure ESM index.js module.

This implements point 3 of the document.