<4> feat: consume LPS-143013-test-2 dependency from LPS-143013-test-1
Do the same as in the previous commit but for sbd package, which is a
cjs2esm bridged package.

Note that we will need to develop something to be able to create the
importmap.json files because it's not obvious where to get imported files
from.

This is because we don't know the exact structure of LPS-143013-test-2 when
building LPS-143013-test-1.

At first sight it may look like adding LPS-143013-test-2 as a dependency of
LPS-143013-test-1 would suffice to know its structure, but it turns out
that the way npm resolves LPS-143013-test-2 dependency tree inside
LPS-143013-test-1 may be different from the way it is resolved as a
standalone project (think of how yarn hoists dependencies too).

We may need to normalize the location of npm packages inside the JAR files
like we did with bundler 2.