<a> feat: demonstrate react based custom element
Render the custom element with portal's copy of React. We simply need to
import 'react' and 'react-dom' because the portal will render an import
map pointing those bare identifiers to the copy of React provided by
frontend-js-react-web.

Also note that we add a npm script called 'serve' that launches a simple
web server to be able to consume the scripts and the CSS file from portal.

In this case I'm using a web server, but we could use webpack's dev server
instead to be able to implement hot reload without any need to deploy
anything. The developer would simply modify the sources in this project,
webpack would rebuild them, and the browser loading the page with the
custom element would simply need to be reloaded for the change to take
effect \o/.