<b> feat: use CommonJS dependency via CJS2ESM bridge
This commit consume `sbd` (a CommonJS package) using a CommonJS to ESM
bridge and providing an importmap.json file.

The build processes the importmap.json file to generate the bridge, then
points the resulting importmap.json to that bridge so that the server takes
on and renders the importmap for the custom element to use it.